requirements.txt
flask
flask-sqlalchemy
flask-restx
pymysql
werkzeug
PyJWT
